Keyguard isn't visible if it hasn't been drawn.
Reduces jank in multiuser lock from QuickSettings. The launcher and
wallpaper were being hidden as soon as the surface for the keyguard
was created. Now they are not hidden until the keyguard has been
drawn. This still leaves a short time where there is a black screen
but it is considerably shorter than it was. Comparable to jb now.

Fixes bug 11046339.

Change-Id: I349d95dba72da27e5c05a7a64c95a2774d17a34e