Fix bug in PathMeasure with trimmed paths
Logic in the hardware renderer optimizes path rendering for simple
paths that consist of only rectangles. Any operation on the path that
adds any other primitive sets the simplicity flag to false appropriately.
PathMeasure.getSegment(), however, avoids those code paths and never sets
the mIsSimple flag at all, so it returns its original value (true, if the
path was simply constructed with no operations before being used in the
getSegment() call).

The fix is to avoid the optimization for paths used in getSegment(), since
it's not clear what operations will end up in the path and it's likely not
going to be just simple rectangles in any case.

Issue #12533902 PathMeasure.getSegment is broken

Change-Id: I71e77207e4d5c649c326557d33eba31e9b0fd45e