Add a warning about the SSID format in CaptivePortalTracker.
WifiInfo#getSSID()'s behavior around surrounding the SSID with double
quotes changed in API 17. Since clients of CaptivePortalTracker may
have already taken this into account, we should make sure not to add
logic which changes the format of the SSID included in latency
broadcasts, soas not to break such logic.

Bug: 12934066
Change-Id: I8800da42c09f429c90409083af1b31c9f720d7d4