Fixed a bug where the current wallpaper could be reset on an update
There is a race condition which caused the wallpaper to be reset
on an app update since the broadcast notifying the service is async.
This CL corrects this behaviour by enforcing that we only reset the
wallpaper if its service was shut down twice in a certain timeframe.
Before, the condition falsely was checking whether the service was
started in the last couple of millis instead of killed.

Bug: 11901821
Change-Id: Icfbc7a5df63215079a83805c5187a3182b192757