Don't wait for finishing animation when second display is disconnected
Basically WindowManagerService wait for finishing animation when
a window is removed. But when second display is disconnected, windows
on second display can't be shown even if animation is waited for.
On the contrary, it keeps on waiting for finishing the animation
in special case.

With this fix windows are immediately removed without waiting for
animation when second display is disconnected.

(Cherry picked from aosp 8e4d505244412facd3630e2c0148a476b48e8bce)

Change-Id: I1354c193c04db394a21a11c174e10c8e7da17a0e

Fixes bug: b/13632129