Back up the preferred home app, if any
If the user has stated a preference about their home app, make sure we
capture that so that a system restore can return them to that preferred
situation.  It's built into the system metadata so that we can, if
necessary, fast-path configuration of that home app while the rest of
the restore operation is in flight.

Change-Id: I64dfee8f7a2a9e40f556cd19298d7b367c6aa8dc