Teach accessibility manager services about user profiles.
For user profiles the accessibility layer will use the accessibility
state of the their owner. A profile cannot have its own plugins or
settings. The rationale behind this is that if a user with impaired
vision is using the device with different profiles linked to his
main one, then he/she should be able to interact with the device no
matter what profile the app is being run with. Hence, The behavior
of the accessibility features should be consistent across apps run
with different profile.

bug:14564491

Change-Id: Id19ce81e1a2dfa48015c7ab3053d66c058be3e46